Add support for cabal >2.0
They changed their internal representation of object keys in a
breaking way.

I opted to leave the API as is, that is still take normale `Text` as
keys and convert to their `Key` type at the last possible moment. This
means this library does not have a breaking change.

I compiled it manually, once with 1.5.6.0 and once with 2.0.3.0.